Split convert.cpp into smaller pieces
